requirements.txt
accelerate
argparse
bitsandbytes
datasets
diffusers
ftfy
gradio
gradio_image_prompter
huggingface_hub
matplotlib
numpy
peft
pillow
regex
scipy
torch
torchvision
tqdm
transformers
wandb